extract out profile
