+++ src/components/Profile/index.ts
@@ -1,0 +1,3 @@
+import Profile from './Profile';
+
+export default Profile;
